+++ CHANGELOG.md
@@ -107,11 +107,11 @@
 - Enh #991: Improve types in `ConnectionInterface::transaction()` (@kikara)
 - Chg #998: Add `yiisoft/db-implementation` virtual package as dependency (@vjik)
 - Chg #999: Remove `requireTransaction()` method and `$isolationLevel` property from `AbstractCommand` (@vjik)
-- Enh #1000: Prepare values in `HashCondition` (@vjik)
+- Enh #1000: Prepare values in `Columns` (@vjik)
 - Chg #1001: Remove `ParamInterface` (@vjik)
 - Chg #1001: Add public properties `$type` and `$value` to `Param` class instead of `getType()` and `getValue()` methods that were removed (@vjik)
 - Chg #1002: Remove specific condition interfaces (@vjik)
-- Chg #1003: Refactor namespace of condition objects and use promoted properties instead of getters (@vjik)
+- Chg #1003, #1006: Refactor namespace of condition objects and use promoted properties instead of getters (@vjik)
 
 ## 1.3.0 March 21, 2024